refactor: lock layer boundaries — move fence parsing to shared + eslint rules (#668)
The four-layer topology (pure shared / main / preload / renderer) was
maintained by convention alone, and one violation had already slipped
through: src/main/compute/save-cell-output.ts imported findRunnableFences,
codeOf, and FenceRange from src/renderer/lib/editor/output-block — main
depending on renderer.

- Move the language-agnostic fence-scanning primitives (findRunnableFences,
  codeOf, FenceRange) into src/shared/compute/fences.ts. Both the main-process
  derived-note writer and the renderer's compute cells now import from shared;
  the renderer-only output-block *editing* helpers (planOutputEdit,
  findAdjacentOutputBlock, resultToJson) stay put. Their tests move to
  tests/shared/compute/fences.test.ts alongside the new home.
- Add type-aware @typescript-eslint/no-restricted-imports boundary rules:
  shared imports nothing from main/renderer/preload/node-builtins; main may
  not import renderer; renderer may not import main. This freezes the
  currently-clean topology so the next stray import fails lint instead of
  review. Verified the rule fires on a probe import.

pnpm lint (0 errors) + full suite (2733 passed) green.
